ctos/drivers/perp-dex-tools-main/para_requirements.txt
python-dotenv>=1.0.0
pytz>=2025.2
asyncio==4.0.0
aiohttp>=3.8.0
websocket-client>=1.6.0
pydantic>=1.8.0
pycryptodome>=3.15.0
ecdsa>=0.17.0
requests==2.32.5
eval_type_backport
tenacity>=9.1.2

# Paradex Python SDK
git+https://github.com/tradeparadex/paradex-py.git@7eb7aa3825d466b2f14abd3e94f2ce6b002d6a63